Tidy all panic/hang/crash code into one file
